NFV-Firewall/Backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
lxml==4.9.3